simulator/requirements.txt
# Core dependencies
requests==2.31.0
kafka-python==2.0.2

# Additional dependencies for data processing
numpy==1.24.3
pandas==2.0.3

# Logging and monitoring
structlog==23.1.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1